Fixed pl_pier, separated pl_pier, pl_subzero_a1 and pl_timbermill_rc4
Fixed an issue with the pl_pier cart (which had an additional burger model attached). Separated pl_pier to a separate config because of that.

Also separated pl_timbermill_rc* and pl_subzero_a*.
No Regex names, because apparently regex filenames are not handled by STT. The names are based on the versions I have used on my server, which should be the latest versions as of now.